vectordb/eval/requirements.txt
absl-py
chromadb
datasets
gcsfs
lancedb
redis
